refactor(cli): split client and todo command modules
